// src/types/reports/reports.ts
export interface IInquiriesListReport {
  name: string;
  email: string;
  subject: string;
  date: string;
  reply: string;
}

export interface IInquiriesListViewModal {
  name: string;
  email: string;
  subject: string;
  address: string;
  city: string;
  country: string;
  state: string;
  zip: string;
  phone: string;
  comment: string;
  date: string;
  status: string;
}
export interface IInquiriesListReplyModalModalProps {
  showModal: boolean;
  setShowModal: (show: boolean) => void;
  modalInformation: IInquiriesListReport | null;
}

export interface IInquiriesListViewModalProps {
  showModal: boolean;
  setShowModal: (show: boolean) => void;
  modalInformation: IInquiriesListReport | null;
}